storage_service: delay deletion of the new_cdc_generation request
We delay the deletion of the new_cdc_generation request to the
moment when the topology transition reaches the
publish_cdc_generation state. We need this change to ensure
adding the CDC generation optimality check in the next commit
has an intended effect. If we didn't make it, it would be possible
that a task makes the new_cdc_generation request, and then, after
this request was removed but before committing the new generation,
another task also makes the new_cdc_generation request. In such
a scenario, two generations are created, but only one should.
After the change introduced by this commit, the second request
would have no effect.