Table: (Mis-)use truncation table to watershed mutations to replay for compacted tables
Fixes #14870 (at least partially)

When doing a compaction with potential GC of tombstones
Put an entry into truncation table, corresponding to the last RP flushed to sstable
(note - does not even have to be related to what we compacted here, only a value
such that we _know_ that data younger that this is already on disk, or potentially
handled by this compation). This will prevent any replay of mutations younger than
this, and ensure nothing we GC:d tombstones for gets resurrected.
TODO: maybe move this to separate table, or rename it to avoid confusion.
NOTE: the timestamp here does not matter, nor does the fact that we don't sync it
across shards, since this is run inside a compaction, i.e. real truncation is
locked out, so said "real" records, where timestamp might matter, will be
done either before or after this. We use the "actual" timestamp as cached in
table object (this will somewhat break if ignore_truncation_record is true, but
so will truncation)